fix: move link saturation to utils
